+++ src/main/java/org/folio/rest/impl/InitAPIs.java
@@ -5,12 +5,16 @@ import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
+import io.vertx.serviceproxy.ServiceBinder;
+import org.folio.rest.resource.interfaces.InitAPI;
+import org.folio.services.PasswordStorageService;
+
 import java.net.URL;
 import java.util.MissingResourceException;
-import org.folio.rest.resource.interfaces.InitAPI;
+
+import static org.folio.util.LoginConfigUtils.PW_CONFIG_PROXY_STORY_ADDRESS;
 
 /**
- *
  * @author kurt
  */
 public class InitAPIs implements InitAPI {
@@ -18,11 +22,14 @@ public class InitAPIs implements InitAPI {
 
   public void init(Vertx vertx, Context context, Handler<AsyncResult<Boolean>> resultHandler) {
     URL u = InitAPIs.class.getClassLoader().getResource(CREDENTIAL_SCHEMA_PATH);
-    if(u == null) {
+    if (u == null) {
       resultHandler.handle(Future.failedFuture(new MissingResourceException(CREDENTIAL_SCHEMA_PATH, InitAPIs.class.getName(), CREDENTIAL_SCHEMA_PATH)));
     } else {
+      new ServiceBinder(vertx)
+        .setAddress(PW_CONFIG_PROXY_STORY_ADDRESS)
+        .register(PasswordStorageService.class, PasswordStorageService.create(vertx));
+
       resultHandler.handle(Future.succeededFuture(true));
     }
   }
-
 }