+++ src/main/java/org/folio/services/package-info.java
@@ -1,0 +1,4 @@
+@ModuleGen(groupPackage = "org.folio.services", name = "login-service")
+package org.folio.services;
+
+import io.vertx.codegen.annotations.ModuleGen;