+++ src/main/java/org/folio/util/LoginConfigUtils.java
@@ -1,0 +1,34 @@
+package org.folio.util;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Future;
+import io.vertx.core.json.JsonObject;
+import org.folio.rest.jaxrs.resource.support.ResponseDelegate;
+
+import javax.ws.rs.core.Response;
+
+public class LoginConfigUtils {
+
+  /**
+   * EvenBus queue
+   */
+  public static final String PW_CONFIG_PROXY_STORY_ADDRESS = "password-store-service.queue";
+
+  /**
+   * Tables
+   */
+  public static final String SNAPSHOTS_TABLE_PW = "auth_password_action";
+  public static final String SNAPSHOTS_TABLE_CREDENTIALS = "auth_credentials";
+
+  private LoginConfigUtils() {
+    //not called
+  }
+
+  public static AsyncResult<Response> createFutureResponse(ResponseDelegate delegate) {
+    return Future.succeededFuture(delegate);
+  }
+
+  public static <T> T getResponseEntity(AsyncResult<JsonObject> asyncResult, Class<T> t) {
+    return asyncResult.result().mapTo(t);
+  }
+}